pyproject.toml
[tool.poetry]
name = "pony-express-rileykraabel"
version = "0.1.0"
description = "This project was created for the University of Utah, CS 4550 course in Spring 2024."
authors = ["Riley Kraabel <rileykraabel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.108.0"
uvicorn = "0.25.0"
pytest = "7.4.0"
httpx = "0.26.0"
sqlmodel = "0.0.16"
passlib = "1.7.4"
python-jose = "3.3.0"
bcrypt = "4.1.2"
cryptography = "42.0.2"
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
